Image
Your homework fulfilled most of the requirements, but you are missing an img tag. Do you need some help with that? or maybe you just forgot :)

Love to see all those section tags
Your naming convention is wise. Naming something " job-description " is exactly the right way.  Naming 2 things "skills_list_1" makes me wonder if "skills-list" would't be a better name, since it's the same class. 
Speaking of that, adding nrs in names is not the best idea. Also using "_" is not that accepted as normal practice.
If you want, you can read more here: https://www.freecodecamp.org/news/css-naming-conventions-that-will-save-you-hours-of-debugging-35cea737d849/

Lastly: On line 34 you have a large text. It would be great if you find a way to write big text, but another way. Maybe you can help me on Sunday and share your findings with us in class. It's something important and good to know, and maybe the rest can learn it also :D